# openapi/v1.yml
swagger: "2.0"
info:
  description: "API definition for Gomematic"
  title: "Gomematic OpenAPI"
  version: "1.0.0-alpha1"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
consumes:
  - "application/json"
produces:
  - "application/json"
schemes:
  - "http"
  - "https"
host: "try.gomematic.tech"
basePath: "/api/v1"
securityDefinitions:
  Basic:
    type: basic
  Header:
    type: apiKey
    in: header
    name: X-API-Key
security:
  - Basic: []
  - Header: []
paths:
...
  /teams:
    get:
      summary: "Fetch all available teams"
      operationId: "ListTeams"
      tags:
        - "team"
      responses:
        200:
          description: "A collection of teams"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/team"
        403:
          description: "User is not authorized"
          schema:
            $ref: "#/definitions/general_error"
        default:
          description: "Some error unrelated to the handler"
          schema:
            $ref: "#/definitions/general_error"

    post:
      summary: "Create a new team"
      operationId: "CreateTeam"
      tags:
        - "team"
      parameters:
        - in: "body"
          name: "team"
          description: "The team data to create"
          required: true
          schema:
            $ref: "#/definitions/team"
            required:
              - "name"
      responses:
        200:
          description: "The created team data"
          schema:
            $ref: "#/definitions/team"
        403:
          description: "User is not authorized"
          schema:
            $ref: "#/definitions/general_error"
        422:
          description: "Failed to validate request"
          schema:
            $ref: "#/definitions/validation_error"
        default:
          description: "Some error unrelated to the handler"
          schema:
            $ref: "#/definitions/general_error"
  /teams/{team_id}:
    get:
      summary: "Fetch a specific team"
      operationId: "ShowTeam"
      tags:
        - "team"
      parameters:
        - in: "path"
          name: "team_id"
          description: "A team UUID or slug"
          type: "string"
          required: true
      responses:
        200:
          description: "The fetched team details"
          schema:
            $ref: "#/definitions/team"
        403:
          description: "User is not authorized"
          schema:
            $ref: "#/definitions/general_error"
        404:
          description: "Team not found"
          schema:
            $ref: "#/definitions/general_error"
        default:
          description: "Some error unrelated to the handler"
          schema:
            $ref: "#/definitions/general_error"

    put:
      summary: "Update a specific team"
      operationId: "UpdateTeam"
      tags:
        - "team"
      parameters:
        - in: "path"
          name: "team_id"
          description: "A team UUID or slug"
          type: "string"
          required: true
        - in: "body"
          name: "team"
          description: "The team data to update"
          required: true
          schema:
            $ref: "#/definitions/team"
      responses:
        200:
          description: "The updated team details"
          schema:
            $ref: "#/definitions/team"
        403:
          description: "User is not authorized"
          schema:
            $ref: "#/definitions/general_error"
        404:
          description: "Team not found"
          schema:
            $ref: "#/definitions/general_error"
        422:
          description: "Failed to validate request"
          schema:
            $ref: "#/definitions/validation_error"
        default:
          description: "Some error unrelated to the handler"
          schema:
            $ref: "#/definitions/general_error"

    delete:
      summary: "Delete a specific team"
      operationId: "DeleteTeam"
      tags:
        - "team"
      parameters:
        - in: "path"
          name: "team_id"
          description: "A team UUID or slug"
          type: "string"
          required: true
      responses:
        200:
          description: "Plain success message"
          schema:
            $ref: "#/definitions/general_error"
        403:
          description: "User is not authorized"
          schema:
            $ref: "#/definitions/general_error"
        404:
          description: "Team not found"
          schema:
            $ref: "#/definitions/general_error"
        400:
          description: "Failed to delete the team"
          schema:
            $ref: "#/definitions/general_error"
        default:
          description: "Some error unrelated to the handler"
          schema:
            $ref: "#/definitions/general_error"
  /teams/{team_id}/users:
    get:
      summary: "Fetch all users assigned to team"
      operationId: "ListTeamUsers"
      tags:
        - "team"
      parameters:
        - in: "path"
          name: "team_id"
          description: "A team UUID or slug"
          type: "string"
          required: true
      responses:
        200:
          description: "A collection of team users"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/team_user"
        403:
          description: "User is not authorized"
          schema:
            $ref: "#/definitions/general_error"
        404:
          description: "Team not found"
          schema:
            $ref: "#/definitions/general_error"
        default:
          description: "Some error unrelated to the handler"
          schema:
            $ref: "#/definitions/general_error"

    post:
      summary: "Assign a user to team"
      operationId: "AppendTeamToUser"
      tags:
        - "team"
      parameters:
        - in: "path"
          name: "team_id"
          description: "A team UUID or slug"
          type: "string"
          required: true
        - in: "body"
          name: "team_user"
          description: "The team user data to assign"
          required: true
          schema:
            $ref: "#/definitions/team_user_params"
      responses:
        200:
          description: "Plain success message"
          schema:
            $ref: "#/definitions/general_error"
        403:
          description: "User is not authorized"
          schema:
            $ref: "#/definitions/general_error"
        404:
          description: "Team or user not found"
          schema:
            $ref: "#/definitions/general_error"
        412:
          description: "User is already assigned"
          schema:
            $ref: "#/definitions/general_error"
        422:
          description: "Failed to validate request"
          schema:
            $ref: "#/definitions/validation_error"
        default:
          description: "Some error unrelated to the handler"
          schema:
            $ref: "#/definitions/general_error"

    put:
      summary: "Update user perms for team"
      operationId: "PermitTeamUser"
      tags:
        - "team"
      parameters:
        - in: "path"
          name: "team_id"
          description: "A team UUID or slug"
          type: "string"
          required: true
        - in: "body"
          name: "team_user"
          description: "The team user data to update"
          required: true
          schema:
            $ref: "#/definitions/team_user_params"
      responses:
        200:
          description: "Plain success message"
          schema:
            $ref: "#/definitions/general_error"
        403:
          description: "User is not authorized"
          schema:
            $ref: "#/definitions/general_error"
        404:
          description: "Team or user not found"
          schema:
            $ref: "#/definitions/general_error"
        412:
          description: "User is not assigned"
          schema:
            $ref: "#/definitions/general_error"
        422:
          description: "Failed to validate request"
          schema:
            $ref: "#/definitions/validation_error"
        default:
          description: "Some error unrelated to the handler"
          schema:
            $ref: "#/definitions/general_error"

    delete:
      summary: "Remove a user from team"
      operationId: "DeleteTeamFromUser"
      tags:
        - "team"
      parameters:
        - in: "path"
          name: "team_id"
          description: "A team UUID or slug"
          type: "string"
          required: true
        - in: "body"
          name: "team_user"
          description: "The team user data to delete"
          required: true
          schema:
            $ref: "#/definitions/team_user_params"
            required:
              - "user_id"
      responses:
        200:
          description: "Plain success message"
          schema:
            $ref: "#/definitions/general_error"
        403:
          description: "User is not authorized"
          schema:
            $ref: "#/definitions/general_error"
        404:
          description: "Team or user not found"
          schema:
            $ref: "#/definitions/general_error"
        412:
          description: "User is not assigned"
          schema:
            $ref: "#/definitions/general_error"
        default:
          description: "Some error unrelated to the handler"
          schema:
            $ref: "#/definitions/general_error"
  /users:
    get:
      summary: "Fetch all available users"
      operationId: "ListUsers"
      tags:
        - "user"
      responses:
        200:
          description: "A collection of users"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/user"
        403:
          description: "User is not authorized"
          schema:
            $ref: "#/definitions/general_error"
        default:
          description: "Some error unrelated to the handler"
          schema:
            $ref: "#/definitions/general_error"

    post:
      summary: "Create a new user"
      operationId: "CreateUser"
      tags:
        - "user"
      parameters:
        - in: "body"
          name: "user"
          description: "The user data to create"
          required: true
          schema:
            $ref: "#/definitions/user"
            required:
              - "username"
              - "password"
              - "email"
      responses:
        200:
          description: "The created user data"
          schema:
            $ref: "#/definitions/user"
        403:
          description: "User is not authorized"
          schema:
            $ref: "#/definitions/general_error"
        422:
          description: "Failed to validate request"
          schema:
            $ref: "#/definitions/validation_error"
        default:
          description: "Some error unrelated to the handler"
          schema:
            $ref: "#/definitions/general_error"
...
definitions:
...
  team:
    type: "object"
    properties:
      id:
        type: "string"
        format: "uuid"
        readOnly: true
      slug:
        type: "string"
        x-omitempty: true
        x-nullable: true
      name:
        type: "string"
        x-omitempty: true
        x-nullable: true
      created_at:
        type: "string"
        format: "date-time"
        readOnly: true
      updated_at:
        type: "string"
        format: "date-time"
        readOnly: true
  team_user:
    type: "object"
    required:
      - "team_id"
      - "user_id"
      - "perm"
    properties:
      team_id:
        type: "string"
        format: "uuid"
      team:
        type: "object"
        $ref: "#/definitions/team"
        readOnly: true
      user_id:
        type: "string"
        format: "uuid"
      user:
        type: "object"
        $ref: "#/definitions/user"
        readOnly: true
      perm:
        type: "string"
        enum:
          - "user"
          - "admin"
          - "owner"
      created_at:
        type: "string"
        format: "date-time"
        readOnly: true
      updated_at:
        type: "string"
        format: "date-time"
        readOnly: true
  user:
    type: "object"
    properties:
      id:
        type: "string"
        format: "uuid"
        readOnly: true
      slug:
        type: "string"
        x-omitempty: true
        x-nullable: true
      username:
        type: "string"
        x-omitempty: true
        x-nullable: true
      password:
        type: "string"
        format: "password"
        x-omitempty: true
        x-nullable: true
      email:
        type: "string"
        x-omitempty: true
        x-nullable: true
      admin:
        type: "boolean"
        x-omitempty: true
        x-nullable: true
      active:
        type: "boolean"
        x-omitempty: true
        x-nullable: true
      created_at:
        type: "string"
        format: "date-time"
        readOnly: true
      updated_at:
        type: "string"
        format: "date-time"
        readOnly: true
  team_user_params:
    type: "object"
    required:
      - "user"
      - "perm"
    properties:
      user:
        type: "string"
      perm:
        type: "string"
        enum:
          - "user"
          - "admin"
          - "owner"
...
  general_error:
    description: General error for regular HTTP status codes
    type: "object"
    required:
      - "status"
      - "message"
    properties:
      status:
        type: "integer"
      message:
        type: "string"
  validation_error:
    description: Validation error which shows failed fields
    type: "object"
    required:
      - "status"
      - "message"
    properties:
      status:
        type: "integer"
      message:
        type: "string"
      errors:
        type: "array"
        items:
          type: "object"
          properties:
            field:
              type: "string"
            message:
              type: "string"